tests: Charset, split tests
