+++ Cargo.toml
@@ -38,3 +38,4 @@ lazy_static = "1.4.0"
 
 [dev-dependencies]
 tiny_http = "^0.7"
+serial_test = "^0.5"
